chore(library): :truck: Move CRC source and header files to `common` subfolder
`CRC.cpp` and `CRC.hpp` will be shared across the Serial Receiver Interface and the Serial Transmitter Interface.
